Add files via upload
Absenteeism at work place. This dataset is provided by UCI Machine Learning Repository. out of which we used this dataset to preprocess and used the preprocessed data to train our model such that it is able to predict any occurence based on the data on which its trained and the factors including in them